Create README.md
This project is not yet complete

         summary of the project
The part of the project you see here is the Admin's
part.This is a part, where staff employed are first 
registered to the database. A form is given to the
new staff which consist of personal details including
there username and passwords, and Access Roles are
being given to them by the Admin according to there 
profession.